+++ CHANGELOG.md
@@ -1,15 +1,16 @@
 ## [Unreleased][unreleased]
 ### Added
+- target, args, kwargs as clearly required arguments
 - get, get_all functions into args_aliases.py
 - processing of different types of values for uid_map and gid_map agrguments
 - dummy functions to Process
-- contexter as required package
 - target_pid, proc, rootdir, workdir arguments into Process.__init__
 - runup, preup, nsenter, chtty, chroot, chdir, networking, postup, preexec, postexec functions into Process
 - proc variable into Clone and Process
 - setns now can take path to namespaces files as arguments
 
 ### Changed
+- add rasing value error in setns
 - processing of args and kwargs in Process.__init__
 - Inject and Chroot classes now based on Container and deprecated